refactor: move BaseModel tests from test_audio.py to test_models.py
