+++ tests/test_models.py
@@ -4,7 +4,13 @@ from src.notebookllama.models import (
     Notebook,
 )
 from src.notebookllama.utils import MindMap, Node, Edge, ClaimVerification
-from src.notebookllama.audio import MultiTurnConversation, ConversationTurn
+from src.notebookllama.audio import (
+    MultiTurnConversation,
+    ConversationTurn,
+    PodcastConfig,
+    VoiceConfig,
+    AudioQuality,
+)
 from pydantic import ValidationError
 
 
@@ -170,3 +176,103 @@ def test_claim_verification() -> None:
             claim_is_true=True,
             supporting_citations=["Support 1", "Support 2", "Support 3", "Support 4"],
         )
+
+
+# Test Audio Configuration Models
+def test_voice_config_defaults():
+    """Test VoiceConfig default values"""
+    config = VoiceConfig()
+    assert config.speaker1_voice_id == "nPczCjzI2devNBz1zQrb"
+    assert config.speaker2_voice_id == "Xb7hH8MSUJpSbSDYk0k2"
+    assert config.model_id == "eleven_turbo_v2_5"
+    assert config.output_format == "mp3_22050_32"
+
+
+def test_voice_config_custom_values():
+    """Test VoiceConfig with custom values"""
+    config = VoiceConfig(
+        speaker1_voice_id="custom_voice_1",
+        speaker2_voice_id="custom_voice_2",
+        model_id="custom_model",
+        output_format="wav_44100_16",
+    )
+    assert config.speaker1_voice_id == "custom_voice_1"
+    assert config.speaker2_voice_id == "custom_voice_2"
+    assert config.model_id == "custom_model"
+    assert config.output_format == "wav_44100_16"
+
+
+def test_audio_quality_defaults():
+    """Test AudioQuality default values"""
+    config = AudioQuality()
+    assert config.bitrate == "320k"
+    assert config.quality_params == ["-q:a", "0"]
+
+
+def test_audio_quality_custom_values():
+    """Test AudioQuality with custom values"""
+    custom_params = ["-q:a", "2", "-compression_level", "5"]
+    config = AudioQuality(bitrate="256k", quality_params=custom_params)
+    assert config.bitrate == "256k"
+    assert config.quality_params == custom_params
+
+
+def test_podcast_config_defaults():
+    """Test that PodcastConfig creates with proper defaults"""
+    config = PodcastConfig()
+
+    assert config.style == "conversational"
+    assert config.tone == "friendly"
+    assert config.focus_topics is None
+    assert config.target_audience == "general"
+    assert config.custom_prompt is None
+    assert config.speaker1_role == "host"
+    assert config.speaker2_role == "guest"
+    assert isinstance(config.voice_config, VoiceConfig)
+    assert isinstance(config.audio_quality, AudioQuality)
+
+
+def test_podcast_config_custom_values():
+    """Test that PodcastConfig accepts custom values"""
+    focus_topics = ["AI Ethics", "Machine Learning", "Future Tech"]
+    custom_prompt = "Make it engaging and technical"
+
+    config = PodcastConfig(
+        style="interview",
+        tone="professional",
+        focus_topics=focus_topics,
+        target_audience="expert",
+        custom_prompt=custom_prompt,
+        speaker1_role="interviewer",
+        speaker2_role="technical_expert",
+    )
+
+    assert config.style == "interview"
+    assert config.tone == "professional"
+    assert config.focus_topics == focus_topics
+    assert config.target_audience == "expert"
+    assert config.custom_prompt == custom_prompt
+    assert config.speaker1_role == "interviewer"
+    assert config.speaker2_role == "technical_expert"
+
+
+def test_podcast_config_validation():
+    """Test that PodcastConfig validates input values"""
+    # Test invalid style
+    with pytest.raises(ValidationError):
+        PodcastConfig(style="invalid_style")
+
+    # Test invalid tone
+    with pytest.raises(ValidationError):
+        PodcastConfig(tone="invalid_tone")
+
+    # Test invalid target_audience
+    with pytest.raises(ValidationError):
+        PodcastConfig(target_audience="invalid_audience")
+
+
+def test_conversation_turn():
+    """Test ConversationTurn model"""
+    turn = ConversationTurn(speaker="speaker1", content="Hello world")
+    assert turn.speaker == "speaker1"
+    assert turn.content == "Hello world"
